ml-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
scikit-learn
numpy